Update bookmarks.md (#256)
Fixed the double-space typo. 

I just copied those two lines as indicated, but the workflow's run jobs kept failing, not realizing that there was extra space. Sadly it took me a couple hours to finally understand what was going on.

It should be updated in case more people struggle with the same issue.